+++ com.unity.cinemachine/Runtime/Core/CinemachineCameraManagerBase.cs
@@ -54,6 +54,7 @@ namespace Cinemachine
         {
             DefaultTarget = default;
             ShowDebugText = false;
+            PriorityAndChannel = OutputChannel.Default;
             InvalidateCameraCache();
         }