+++ com.unity.cinemachine/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - New icons for cinemachine (cameras, components, extensions, tracks).
 - Freelook ForcePosition is more precise now.
 - Confiner2D supports BoxCollider2D now.
+- Added Channel setting, to use for multiple CM Brains, instead of piggybacking on the layer system.
 
 
 ## [3.0.0-pre.1] - 2022-06-01